Add files via upload
Plot showing the adiabaticity parameter (lower plot) and flip probability (upper plot) as a function of distance with 3 different energies. 
The important part is to observe the flip probability, as it tells us if strong flavor conversion occurs or if the evolution is non-adiabatic and that a flip in the mass eigenstate occurs at a resonance. 
This plot was specifically done using the data given.